ci: run on `pull_request_target`
since forks aren't allowed to access secrets running on `pull_request` makes it fail. Making this change should not effected much since everyone works on forks and open a PR against this repo. If ever someone opens a PR from a branch that exists in this repo (which is very unlikley scenario) we can manually check or run danger locally